requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==3.2.15
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
pycparser==2.21
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
